Module Dune_release.Config

type t = {

user : string option;

remote : string;

local : Fpath.t;

keep_v : bool option;

auto_open : bool option;

}

module Opam_repo_fork : sig ... end

val create : ?⁠pkgs:Pkg.t list -> unit -> (unit, Bos_setup.R.msg) Stdlib.result

module Cli : sig ... end

val token : token:string Cli.t option -> dry_run:bool -> unit -> (string, Bos_setup.R.msg) Stdlib.result

Returns the token value that should be used for github API requests. If a token was provided via the CLI, it is returned. Otherwise the token file in the config dir is looked up. If it exists, its content is returned, if it does not, the user is prompted for a token which will be then saved to that file. When dry_run is true it always returns Ok "${token}" but still looks up the relevant config file as it would normally have.

val keep_v : keep_v:bool Cli.t -> (bool, Bos_setup.R.msg) Stdlib.result

val auto_open : no_auto_open:bool Cli.t -> (bool, Bos_setup.R.msg) Stdlib.result

val opam_repo_fork : ?⁠pkgs:Pkg.t list -> remote:string Cli.t option -> local:Fpath.t Cli.t option -> unit -> (Opam_repo_fork.t, Bos_setup.R.msg) Stdlib.result

Returns the opam-repository fork to use, based on the CLI provided values remote and local and the user's configuration.

If both remote and local are provided, they are returned without reading any local configuration.

If either or both of them are None, the configuration is looked up. If it doesn't exist, the interactive creation quizz is started. The configuration values are used to fill up the blanks in remote and local.

pkgs is only used to offer suggestions to the user during the creation quizz.

val load : unit -> (t option, Bos_setup.R.msg) Stdlib.result

val save : t -> (unit, Bos_setup.R.msg) Stdlib.result

val pretty_fields : t -> (string * string option) list

pretty_fields t returns the list of pretty-printed key-value pairs for the config t.

module type S = sig ... end

module Draft_release : S

module Draft_pr : S

module Release_asset_name : S