Module Deprecate.Config_user

val option_doc : string

Documentation bit indicating the --user option is deprecated because it is redundant with the --remote-repo option.

val option_use : string

Message warning users they used the deprecated --user option and that they should use --remote-repo only.

val config_field_doc : string

Documentation bit indicating the user configuration field is deprecated because it is redundant with the remote field.

val config_field_use : string

Message warning users they are setting the deprecated user field of their configuration.