Module Dune_release.Github

module Parse : sig ... end

Publish

val publish_distrib : token:string -> dry_run:bool -> msg:string -> archive:Fpath.t -> yes:bool -> draft:bool -> Pkg.t -> (string, Bos_setup.R.msg) Result.result

Push the tag, create a Github release, upload the distribution archive and return the release archive download URL

val publish_doc : dry_run:bool -> msg:string -> docdir:Fpath.t -> yes:bool -> Pkg.t -> (unit, Bos_setup.R.msg) Result.result

val undraft_release : token:string -> dry_run:bool -> owner:string -> repo:string -> release_id:string -> name:string -> (string, Bos_setup.R.msg) Result.result

undraft_release updates an existing release to undraft it and returns the release archive download URL.

val open_pr : token:string -> dry_run:bool -> title:string -> fork_owner:string -> branch:Vcs.commit_ish -> opam_repo:(string * string) -> draft:bool -> string -> Pkg.t -> ([ `Url of string | `Already_exists ], Bos_setup.R.msg) Bos_setup.result

val undraft_pr : token:string -> dry_run:bool -> opam_repo:(string * string) -> pr_id:string -> (string, Bos_setup.R.msg) Bos_setup.result

undraft_pr updates an existing pull request to undraft it and returns the pull request URL.