Module Github_v3_api.Pull_request_response

val html_url : Yojson.Basic.t -> ([ `Already_exists | `Url of string ], Bos_setup.R.msg) Bos_setup.result

html_url response extracts the html_url field from a github json response, or `Already_exists if the corresponding pull request already exists, or error messages.