Module Dune_release__Distrib

Distribution description

Distribution description

type t

The type for describing distribution creation.

val v : ?⁠massage:(unit -> (unit, Rresult.R.msg) Rresult.result) -> ?⁠exclude_paths:(unit -> (Fpath.t list, Rresult.R.msg) Rresult.result) -> unit -> t

distrib ~massage ~exclude_paths () influences the distribution creation process performed by the dune-release tool. See the full details about distribution creation.

In the following the distribution build directory is a private clone of the package's source repository's HEAD when dune-release distrib is invoked.

massage is invoked in the distribution build directory, before archiving. It can be used to generate distribution time build artefacts. Defaults to massage.

exclude_paths () is invoked in the distribution build directory, after massaging, to determine the paths that are excluded from being added to the distribution archive. Defaults to exclude_paths.

val massage : t -> unit -> (unit, Rresult.R.msg) Rresult.result

val exclude_paths : t -> unit -> (Fpath.t list, Rresult.R.msg) Rresult.result