Module Archive.Response

val browser_download_url : Yojson.Basic.t -> (string, Bos_setup.R.msg) Bos_setup.result

browser_download_url response extracts the browser_download_url field from a github release asset upload response, or error messages.

val name : Yojson.Basic.t -> (string, Bos_setup.R.msg) Bos_setup.result

name response extracts the github name for the asset, which might differ from the filename for the archive we uploaded.