Module Release.Response

val browser_download_url : name:string -> Yojson.Basic.t -> (string, Bos_setup.R.msg) Bos_setup.result

browser_download_url ~release_id response extracts the browser_download_url field from a github release asset upload response named name, or error messages.

val release_id : Yojson.Basic.t -> (int, Bos_setup.R.msg) Bos_setup.result

release_id response extracts the id field from a github response, or error messages.