Module Dune_release__Json

val from_string : string -> (Yojson.Basic.t, Bos_setup.R.msg) Bos_setup.result

from_string s parses s and builds a Yojson.Basic.t type accordingly, or returns the associated error message if the input is not a valid JSON value.

val string_field : field:string -> Yojson.Basic.t -> (string, Bos_setup.R.msg) Bos_setup.result

string_field ~field j returns the value of field field from the JSON value j if it is a string, or returns the associated error message otherwise.

val int_field : field:string -> Yojson.Basic.t -> (int, Bos_setup.R.msg) Bos_setup.result

int_field ~field j returns the value of field field from the JSON value j if it is an integer, or returns the associated error message otherwise.

val list_field : field:string -> Yojson.Basic.t -> (Yojson.Basic.t list, Bos_setup.R.msg) Bos_setup.result

list_field ~field j returns the list of values of field field from the JSON value j if it is a list, or returns the associated error message otherwise.