Module Dune_release__Opam.File

opam files

opam file

val fields : dry_run:bool -> Fpath.t -> (string list Bos_setup.String.map, Bos_setup.R.msg) Bos_setup.result

fields f returns a simplified model of the fields of the opam file f. The domain of the result is included in field_names. Note that the depends: and depopts: fields are returned without version constraints.