Module Dune_release__Stdext.Path

val is_backup_file : string -> bool

is_backup_file s returns true iff the filename s:

ends with '~'

or begins with '#' and ends with '#'.

val find_files : names_wo_ext:string list -> Fpath.t list -> Fpath.t list

find_files ~names_wo_ext files returns the list of files among files whose name without extension is equal to an element of names_wo_ext. Backup files are ignored.